db/requirements.txt
mysql-connector-python
urllib3
rich
requests
pyyaml
dotenv